// MochiR.Client/src/app/api/models/search-result-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SearchResultType } from '../models/search-result-type';
export interface SearchResultDto {
  createdAtUtc: string;
  excerpt: string | null;
  reviewId: number | null;
  score: number;
  subjectId: number | null;
  subtitle: string | null;
  title: string;
  type: SearchResultType;
}